Group extensions into new files
