+++ macOS/Extensions/Cocoa+Extensions.swift
@@ -1,5 +1,5 @@
 //
-//  Extensions.swift
+//  Cocoa+Extensions.swift
 //  Accelerate
 //
 //  Created by Ritam Sarmah on 9/3/19.
@@ -7,51 +7,8 @@
 //
 
 import Carbon
+import Cocoa
 import Defaults
-import Foundation
-import SafariServices
-
-extension SFSafariPage {
-    func triggerAction(for shortcut: Shortcut) {
-        dispatchMessageToScript(withName: "triggerAction", userInfo: ["shortcut": shortcut.dictionaryRepresentation])
-    }
-
-    func isAllowed(completion: @escaping (_ isAllowed: Bool, _ rule: String?) -> Void) {
-        getPropertiesWithCompletionHandler { properties in
-            guard let url = properties?.url else {
-                completion(false, nil)
-                return
-            }
-
-            let blocklistRule = self.findRule(for: url, in: Defaults[.blocklist])
-            let isBlocklistInverted = Defaults[.isBlocklistInverted]
-            let isPageAllowed = blocklistRule != nil ? isBlocklistInverted : !isBlocklistInverted
-
-            completion(isPageAllowed, blocklistRule)
-        }
-    }
-
-    private func findRule(for url: URL, in rules: [String]) -> String? {
-        let prefix = "^(https?://)?(www\\.)?"
-        let urlString = url.absoluteString.replacingFirstOccurrence(of: prefix, with: "", options: .regularExpression)
-
-        return
-            rules
-            .filter { !$0.isEmpty }
-            .first { rule in
-                let predicateRule =
-                    rule
-                    .replacingOccurrences(of: "?", with: "\\?")  // Escape question marks in URL
-                    .replacingFirstOccurrence(of: prefix, with: "", options: .regularExpression)
-
-                // [c] means case insensitive
-                let predicate = NSPredicate(format: "SELF LIKE[c] %@", "\(predicateRule)*")
-
-                // Return whether URL matches a rule
-                return predicate.evaluate(with: urlString)
-            }
-    }
-}
 
 extension NSEvent.ModifierFlags: Codable {
     public init(from decoder: Decoder) throws {
@@ -147,13 +104,6 @@ extension NSViewController {
     }
 }
 
-extension [Shortcut] {
-    var toolbarShortcutIndex: Int? {
-        guard let toolbarShortcutIdentifier = Defaults[.toolbarShortcutIdentifier] else { return nil }
-        return Defaults[.shortcuts].firstIndex(where: { $0.identifier == toolbarShortcutIdentifier })
-    }
-}
-
 extension NSAlert {
     static func showAlert(title: String, message: String?) {
         DispatchQueue.main.async {
@@ -243,13 +193,6 @@ extension NSButton {
     func setCheckboxState(with value: Bool) {
         self.state = value ? .on : .off
     }
-}
-
-// Custom nil coalescing operator for providing a default String value to any optional type
-infix operator ??? : NilCoalescingPrecedence
-
-public func ??? (optional: (some Any)?, defaultValue: @autoclosure () -> String) -> String {
-    optional.map { String(describing: $0) } ?? defaultValue()
 }
 
 extension NSTableView {
