+++ macOS/Extensions/Defaults+Keys.swift
@@ -21,3 +21,10 @@ extension Defaults.Keys {
     // Array of shortcut preference keys
     static let allShortcutKeys: [Defaults.AnyKey] = [.shortcuts, .toolbarShortcutIdentifier]
 }
+
+extension [Shortcut] {
+    var toolbarShortcutIndex: Int? {
+        guard let toolbarShortcutIdentifier = Defaults[.toolbarShortcutIdentifier] else { return nil }
+        return Defaults[.shortcuts].firstIndex(where: { $0.identifier == toolbarShortcutIdentifier })
+    }
+}
